Cleanup the metrics and statsd middlewares.
* Get rid of metrics.report. Update tests to use listened for messages.
* Move logging middleware related functionality to logging/middleware
* formatting
