+++ bin/router
@@ -15,7 +15,7 @@ urlparse = require('urlparse'),
 express = require('express');
 
 const
-metrics = require('../lib/logging/metrics.js'),
+metrics = require('../lib/logging/middleware/metrics.js'),
 wsapi = require('../lib/wsapi.js'),
 config = require('../lib/configuration.js'),
 heartbeat = require('../lib/heartbeat.js'),
@@ -24,7 +24,7 @@ forward = require('../lib/http_forward').forward,
 addP3PHeader = require('../lib/p3p.js'),
 shutdown = require('../lib/shutdown.js'),
 toobusy = require('../lib/busy_middleware.js'),
-statsd = require('../lib/logging/statsd.js');
+statsd = require('../lib/logging/middleware/statsd.js');
 
 var app = express.createServer();
 
@@ -150,7 +150,7 @@ wsapi.routeSetup(app, {
   write_url: dbwriter_url
 });
 
-app.use(metrics.metricsMiddleware);
+app.use(metrics);
 
 //catch-all
 app.use(function(req, res, next) {
