+++ lib/logging/middleware/metrics.js
@@ -6,12 +6,12 @@
  * Metrics middleware that reports when the signin dialog is opened.
  */
 
-const logger = require("./logging").logger;
+const logger = require("../logging").logger;
 const urlparse = require('urlparse');
 
 
 // utility function to log a bunch of stuff at user entry point
-exports.metricsMiddleware = function(req, res, next) {
+module.exports = function(req, res, next) {
   if (req.url === '/sign_in') userEntry(req);
   next();
 };
