+++ lib/logging/transports/filters/metrics.js
@@ -38,11 +38,13 @@ exports.toEntry = function(msg, entry) {
 
   // allow convenient reporting of atoms by converting atoms into objects
   if (entry === null || typeof entry !== 'object') entry = { msg: entry };
-  if (entry.type) throw "reported metrics may not have a `type` property, that's reserved";
+  if (entry.type)
+    throw "reported metrics may not have a `type` property, that's reserved";
   entry.type = exports.toType(msg);
 
   // timestamp
-  if (entry.at) throw "reported metrics may not have an `at` property, that's reserved";
+  if (entry.at)
+    throw "reported metrics may not have an `at` property, that's reserved";
   entry.at = new Date().toUTCString();
 
   return entry;
