+++ lib/logging/transports/metrics-kpi.js
@@ -23,6 +23,7 @@ MetricsKpiggybankTransport.prototype.log
   if ( ! filter.test(msg)) return callback(null, true);
   if ( ! config.get('kpi_send_metrics')) return callback(null, true);
 
+
   var entry = filter.toEntry(msg, meta);
   this.queue.push(entry);
