+++ lib/wsapi.js
@@ -27,7 +27,6 @@ fs = require('fs'),
 path = require('path'),
 validate = require('./validate'),
 version = require('./version.js'),
-statsd = require('./logging/statsd'),
 bcrypt = require('./bcrypt'),
 i18n = require('i18n-abide'),
 i18n_check = require('./i18n_client_check'),
